// packages/vite/types.ts
interface NodeRedLauncherOptions {
  runtime?: {
    port?: number;
    settingsFilepath?: string;
    version?: string;
  };
  restartDelay?: number;
  args?: string[];
}